+++ ChangeLog.md
@@ -13,8 +13,10 @@
   - Entferne Modal-Zustand. Verwende stattdessen Nachrichten & Konstruktor-Parameter.
   - Füge padding beim bewegen-Widget hinzu.
   - Zugkontrolle-Parameter `S` muss zusätzlich `Display` implementieren.
+  - Anpassen der `Nachricht`-Typen.
 - Dunkles Thema. Dazu wurden u.a. die `zeichnen` Methode des `Cache` angepasst.
 - Verwende PickList für Modus-Auswahl.
+- Verwende `rfd` zur Datei-Auswahl beim Speichern und Laden.
 - Anpassen der Widget-Anordnung der Kopf-Leiste.
 - Verbesserungen im `Gleise`-canvas:
   - Touch-Bewegungen aktualisieren nicht mehr die gespeicherte Maus-Position.